third_party/WebKit/LayoutTests/fast/css/quotes-parsing.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="div"></div>
<script>
test(function(){
    div.style.quotes = "red";
    assert_equals(div.style.quotes, "");
    div.style.quotes = '"a"';
    assert_equals(div.style.quotes, "");
    div.style.quotes = '"a" "b" cake';
    assert_equals(div.style.quotes, "");
}, "Invalid strings for quotes property do not parse");

test(function(){
    div.style.quotes = '"a" "b"';
    assert_equals(div.style.quotes, '"a" "b"');
}, "Valid strings should be serialized with space separator");
</script>
